Contactors/Prechargers Driver Fixes (#187)
* Contactors and Prechargers fixes

* made contactors driver "friendly"

* "friendly" precharge drivers

* merge precharge and contactor

* fix compile errors related to merging contactor/precharge driver

* separate pin number and pin bitmask #defines per clark's suggestions

* update contactors_set() call with new enum definitions

* pr fixes: remove test files, pin mask fix for bsp_gpio_read_pin

* sid

+++ BSP/Inc/BSP_GPIO.h
@@ -45,7 +45,7 @@ void BSP_GPIO_Write(port_t port, uint16_t data);
  * @param   pin The pin to read from 
  * @return  State of the pin
  */ 
-uint8_t BSP_GPIO_Read_Pin(port_t port, uint8_t pin);
+uint8_t BSP_GPIO_Read_Pin(port_t port, uint16_t pin);
 
 /**
  * @brief   Writes data to a specified pin
